fix: postgres list_bodies panic and vault-scope the summaries fallback
The postgres list_bodies deserialized INTEGER columns (protocol_version,
suite_id) as i16 while every other query uses i32, so the embed=bodies
listing panicked and returned 503. The frontend silently fell back to
the summaries listing, whose DTOs carry no vault_id — combined with the
new vault scoping this emptied the listing, skipped verifier checks, and
let a wrong master password 'unlock' an empty vault (broke CI).

- postgres: read protocol_version/suite_id as i32, matching get()
- records.js: scope the summaries fallback after the full-record GETs;
  verifyRootKey now walks verifier-then-live candidates, skipping any
  whose full DTO belongs to a foreign vault, so a leftover foreign
  verifier can never lock the vault out
- unit test: summaries fallback with a foreign verifier listed first

+++ backend/crates/postgres/src/lib.rs
@@ -265,8 +265,8 @@ impl OpaqueStore for PostgresStore {
         Ok(rows
             .iter()
             .map(|row| GenericEncryptedRecordV1 {
-                protocol_version: row.get::<_, i16>("protocol_version") as u16,
-                suite_id: row.get::<_, i16>("suite_id") as u16,
+                protocol_version: row.get::<_, i32>("protocol_version") as u16,
+                suite_id: row.get::<_, i32>("suite_id") as u16,
                 deployment_id: row.get("deployment_id"),
                 vault_id: row.get("vault_id"),
                 record_id: row.get("record_id"),
